Provide linkedInteractives in initMessage in the authoring mode
[#172781754]
