+++ app/models/base_interactive.rb
@@ -11,6 +11,27 @@ module BaseInteractive
     end
   end
 
+  # used for react-based authoring
+  def to_authoring_hash
+    hash = to_hash
+    hash[:id] = id
+    hash[:linked_interactive_id] = linked_interactive_id
+    hash[:linked_interactive_type] = linked_interactive_type
+    hash[:aspect_ratio] = aspect_ratio
+    hash[:interactive_item_id] = interactive_item_id
+    hash[:linked_interactive_item_id] = linked_interactive_item_id
+    # Note that linked_interactives is independent from linked_interactive_id and linked_interactive_type fields
+    hash[:linked_interactives] = linked_interactives_list
+    hash
+  end
+
+  def authoring_api_urls(protocol, host)
+    {
+      get_interactive_list: interactive_page ? Rails.application.routes.url_helpers.api_v1_get_interactive_list_url(id: interactive_page.id, protocol: protocol, host: host) : nil,
+      set_linked_interactives: interactive_page ? Rails.application.routes.url_helpers.api_v1_set_linked_interactives_url(id: interactive_page.id, protocol: protocol, host: host) : nil
+    }
+  end
+
   def portal_hash
     result = report_service_hash
     # When Portal receives iframe_interactive, it expects:
