+++ spec/models/mw_interactive_spec.rb
@@ -63,6 +63,7 @@ describe MwInteractive do
       expected[:aspect_ratio] = interactive.aspect_ratio
       expected[:interactive_item_id] = interactive.interactive_item_id
       expected[:linked_interactive_item_id] = interactive.linked_interactive_item_id
+      expected[:linked_interactives] = interactive.linked_interactives_list
       expect(interactive.to_authoring_hash).to eq(expected)
     end
   end
